config/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework import permissions

from drf_yasg.views import get_schema_view

from drf_yasg import openapi

from config import settings

schema_view = get_schema_view(

    openapi.Info(

        title="My API",

        default_version='v1',

        description="Test description",

    ),

    public=True,

    permission_classes=(permissions.AllowAny,),

)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('swagger/', schema_view.with_ui('swagger',
                                         cache_timeout=0), name='schema-swagger-ui'),

    path('redoc/', schema_view.with_ui('redoc', cache_timeout=0),
         name='schema-redoc'),
    path('api/v1/', include(
        [
            path('account/', include('account.urls')),
            path('project/', include('project.urls')),
        ]
    ))
]
if settings.DEBUG:

    import debug_toolbar

    urlpatterns += [

        path('__debug__/', include(debug_toolbar.urls)),

    ]