requirements.txt
djangorestframework==3.16.1
psycopg2-binary==2.9.10
